+++ R/precip_indices.R
@@ -21,7 +21,6 @@
 #' @references 
 #' 
 #' Aguilar E., et al. (2005). Journal of Geophysical Research, 110(D23), D23107.
-#'  \cr\doi{10.1029/2005JD006119}
 #' 
 #' Kehel Z., et al. (2016). In: Applied Mathematics and Omics to Assess Crop
 #'  Genetic Resources for Climate Change Adaptive Traits (eds Bari A., Damania
@@ -33,7 +32,7 @@
 #' 
 #' dates <- c("2017-12-15", "2017-12-31")
 #' 
-#' dt <- get_chirps(lonlat, dates)
+#' dt <- get_chirps(lonlat, dates, server = "ClimateSERV")
 #' 
 #' # take the indices for the entire period
 #' precip_indices(dt, timeseries = FALSE)